// generated/models/ApiCompetitionRulesInput.ts
// @ts-nocheck
/**
 * 6529.io API
 * JSON REST API for 6529.io. New here? A terminology overview and a step-by-step authentication walkthrough (guides &amp; auth) live at <a href=\"https://6529.io/tools/api\">https://6529.io/tools/api</a>. The raw machine-readable spec is downloadable at <a href=\"/openapi.yaml\">/openapi.yaml</a> and <a href=\"/openapi.json\">/openapi.json</a>.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ApiWaveDecisionsStrategy } from '../models/ApiWaveDecisionsStrategy';
import { HttpFile } from '../http/http';

export class ApiCompetitionRulesInput {
    'type': ApiCompetitionRulesInputTypeEnum;
    'winning_threshold': number | null;
    'winning_threshold_min_duration_ms': number | null;
    'max_winners': number | null;
    'max_votes_per_identity_to_drop': number | null;
    'time_lock_ms': number | null;
    'decisions_strategy': ApiWaveDecisionsStrategy | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "type",
            "baseName": "type",
            "type": "ApiCompetitionRulesInputTypeEnum",
            "format": ""
        },
        {
            "name": "winning_threshold",
            "baseName": "winning_threshold",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "winning_threshold_min_duration_ms",
            "baseName": "winning_threshold_min_duration_ms",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "max_winners",
            "baseName": "max_winners",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "max_votes_per_identity_to_drop",
            "baseName": "max_votes_per_identity_to_drop",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "time_lock_ms",
            "baseName": "time_lock_ms",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "decisions_strategy",
            "baseName": "decisions_strategy",
            "type": "ApiWaveDecisionsStrategy",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ApiCompetitionRulesInput.attributeTypeMap;
    }

    public constructor() {
    }
}

export enum ApiCompetitionRulesInputTypeEnum {
    Rank = 'RANK',
    Approve = 'APPROVE'
}